Support loading templates from module importer
In modules that require files from the *importer* such as those using
`{{ code('.....') }}` the fact that the parent of the module is not
included in the file search path causes it to be impossible to do things
like have a `lambda-python` module that requires the caller to provide a
`handler.py` to inline to the function. This would be extra useful for a
custom-resource module.