Remove indents for templates near the max length
Indenting 4 characters can make templates too big for the normal
CloudFormation size limit. One example I have is a template that
(without indents) is 20684 characters expands to 51312 characters with
the formica 4-char default indentation. For reference, the YAML
representation of the same template is 24082, so based on my sample of 1
template the most efficient representation is non-indented JSON.

This change adds a check when returning a JSONified template that it's
under the CloudFormation length limit.